Did more model validation
I changed the random constant to 100. This just means i tried 100 different arangements of validation and train data for each estimator. I tried a couple of different states for the original train test split. In the end I found that around 75 estimators performed similarly to 1000 or more estimators on the validation set, while not being as overfit. Fewer than 75 often compromised the validation error result.