As predicted earlier, this model isn't generalizable to the rest of the year
Is this within the scope of this project to handle? Yeah kinda sorta. But i also could leave as is. I am not even sure quite how i might do this though... My first thought is Neural Networks to maybe learn the importance of day of the week but im not sure that will improve the performance much.